FEAT: Docking station, can add generators to the modular.

+++ game/module_cannon.cpp
@@ -27,7 +27,7 @@ void mark::module::cannon::tick(mark::tick_context& context) {
 				enemy->damage(1, cur);
 			}
 			for (int i = 0; i < 4; i++) {
-				const auto velocity = socket()->world().resource_manager().random_double(25, 50);
+				const auto velocity = static_cast<float>(socket()->world().resource_manager().random_double(25, 50));
 				const auto direction = static_cast<float>(socket()->world().resource_manager().random_double(90, 270)) + rotation;
 				context.particles.push_back(mark::particle(m_im_ray, cur, velocity, direction, 1.f, { 255, 0, 0 }));
 			}
