+++ game/module_shield_generator.cpp
@@ -11,8 +11,8 @@ mark::module::shield_generator::shield_generator(mark::resource::manager& resour
 	m_im_generator(resource_manager.image("shield-generator.png")),
 	m_im_shield(resource_manager.image("shield.png")),
 	m_im_glare(resource_manager.image("glare.png")),
-	m_light_lfo(1.0, resource_manager.random_double(0, 1)),
-	m_shield_lfo(1.1, resource_manager.random_double(0, 1)) {
+	m_light_lfo(1.f, static_cast<float>(resource_manager.random_double(0, 1))),
+	m_shield_lfo(1.1f, static_cast<float>(resource_manager.random_double(0, 1))) {
 
 }
 
@@ -20,8 +20,8 @@ void mark::module::shield_generator::tick(mark::tick_context& context) {
 	m_shield_lfo.tick(context.dt);
 	m_light_lfo.tick(context.dt);
 	const auto pos = socket()->relative_pos();
-	const auto light_intensity = 255.f * (m_light_lfo.get() + 1.f) / 2.f;
-	const auto shield_intensity = 255.f * (m_shield_lfo.get() + 1.f) / 2.f;
+	const auto light_intensity = static_cast<uint8_t>(255.f * (m_light_lfo.get() + 1.f) / 2.f);
+	const auto shield_intensity = static_cast<uint8_t>(255.f * (m_shield_lfo.get() + 1.f) / 2.f);
 	context.sprites[0].push_back(mark::sprite(m_im_generator, pos, 32.f, socket()->rotation()));
 	context.sprites[1].push_back(mark::sprite(m_im_shield, pos, 128.f, 0.0f, 0, sf::Color(255, 255, 255, shield_intensity)));
 	context.sprites[2].push_back(mark::sprite(m_im_glare, pos, 64.f, 0.0f, 0, sf::Color(0, 255, 255, light_intensity)));
